adds peaches
adds the basic code and sprites for the hydroponics plant peach, its tree and fruit and seed. Does not put it in the game anywhere or use it for anything, those will be in a later commit. Just the peach and the theoretical ability to grow one if we somehow get a seed.